Add `git cherry-pick --mainline-auto` option
Usually you cannot cherry-pick a merge because you do not know which
side of the merge should be considered the mainline. You can use
`--mainline` to specify the parent index.

In some situations, however, it's "obvious" which mainline to use,
because only one parent is a non-merge commit.

For that situation, this option will do the cherry-pick if one (and only
one) merge parent is not a merge. If both `-a` and `-m <number>` are
specified, `--mainline` takes precedence.

Using this option can be useful when the branch from which you are
cherry-picking typically only gets new commits merged in as merge
commits, whether dictated by contributor guidelines or enforced by a git
hosting service.